+++ src/api/user_data.ts
@@ -92,6 +92,16 @@ export default class UserData {
     //      Class Methods     //
     ////////////////////////////
 
+    // Return an array of keywords the user has muted on the server side
+    static async mutedKeywords(): Promise<string[]> {
+        const serverSideFilters = await MastoApi.instance.getServerSideFilters();
+        console.log(`mutedKeywords() raw serverSideFilters:`, serverSideFilters);
+        let keywords = serverSideFilters.map(f => f.keywords.map(k => k.keyword)).flat().flat().flat();
+        keywords = keywords.map(k => k.toLowerCase().replace(/^#/, ""));
+        console.log(`mutedKeywords() found ${keywords.length} keywords:`, keywords);
+        return keywords;
+    }
+
     // Fetch or load array of TrendingTags sorted by number of times the user tooted it
     static async getPostedHashtagsSorted(): Promise<TrendingTag[]> {
         const userTags = await UserData.getPostedHashtags();